backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
openai>=1.14.0
anthropic>=0.19.0
playwright>=1.42.0
beautifulsoup4>=4.12.3
httpx>=0.27.0
sentry-sdk>=1.40.0
supabase>=2.4.0
redis>=5.0.1
python-dotenv>=1.0.1
aiofiles>=23.2.1
email-validator>=2.1.1
jinja2>=3.1.3
numpy>=1.26.0
